Small cleanups.  Move reload functions to DocumentCommandHandlers.
