+++ src/command/Commands.js
@@ -132,8 +132,11 @@ define(function (require, exports, module) {
     exports.HELP_ABOUT                  = "help.about";                 // HelpCommandHandlers.js       _handleAboutDialog()
 
     // APP
+    exports.APP_RELOAD                  = "app.reload";                 // DocumentCommandHandlers.js   handleReload()
+    exports.APP_RELOAD_WITHOUT_EXTS     = "app.reload_without_exts";    // DocumentCommandHandlers.js   handleReloadWithoutExts()
+    
     // File shell callbacks - string must MATCH string in native code (appshell/command_callbacks.h)
-    exports.APP_ABORT_QUIT              = "app.abort_quit";             // DocumentCommandHandlers.js   _handleAbortQuit()
-    exports.APP_BEFORE_MENUPOPUP        = "app.before_menupopup";       // DocumentCommandHandlers.js   _handleBeforeMenuPopup()
+    exports.APP_ABORT_QUIT              = "app.abort_quit";             // DocumentCommandHandlers.js   handleAbortQuit()
+    exports.APP_BEFORE_MENUPOPUP        = "app.before_menupopup";       // DocumentCommandHandlers.js   handleBeforeMenuPopup()
 });
