+++ src/document/DocumentCommandHandlers.js
@@ -23,10 +23,12 @@
 
 
 /*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50, regexp: true */
-/*global define, $, brackets, window */
+/*global define, $, brackets, window, WebSocket */
 
 define(function (require, exports, module) {
     "use strict";
+    
+    var _ = require("thirdparty/lodash");
     
     // Load dependent modules
     var AppInit             = require("utils/AppInit"),
@@ -50,7 +52,11 @@ define(function (require, exports, module) {
         DragAndDrop         = require("utils/DragAndDrop"),
         PerfUtils           = require("utils/PerfUtils"),
         KeyEvent            = require("utils/KeyEvent"),
-        LanguageManager     = require("language/LanguageManager");
+        LanguageManager     = require("language/LanguageManager"),
+        Inspector           = require("LiveDevelopment/Inspector/Inspector"),
+        Menus               = require("command/Menus"),
+        UrlParams           = require("utils/UrlParams").UrlParams,
+        StatusBar           = require("widgets/StatusBar");
     
     /**
      * Handlers for commands related to document handling (opening, saving, etc.)
@@ -1221,7 +1227,7 @@ define(function (require, exports, module) {
      * @private
      * Implementation for abortQuit callback to reset quit sequence settings
      */
-    function _handleAbortQuit() {
+    function handleAbortQuit() {
         _windowGoingAway = false;
     }
     
@@ -1229,7 +1235,7 @@ define(function (require, exports, module) {
      * @private
      * Implementation for native APP_BEFORE_MENUPOPUP callback to trigger beforeMenuPopup event
      */
-    function _handleBeforeMenuPopup() {
+    function handleBeforeMenuPopup() {
         $(PopUpManager).triggerHandler("beforeMenuPopup");
     }
     
@@ -1372,13 +1378,143 @@ define(function (require, exports, module) {
         }
     }
     
-    // Init DOM elements
+    /**
+     * Disables Brackets' cache via the remote debugging protocol.
+     * @return {$.Promise} A jQuery promise that will be resolved when the cache is disabled and be rejected in any other case
+     */
+    function _disableCache() {
+        var result = new $.Deferred();
+        
+        if (brackets.inBrowser) {
+            result.resolve();
+        } else {
+            var port = brackets.app.getRemoteDebuggingPort ? brackets.app.getRemoteDebuggingPort() : 9234;
+            Inspector.getDebuggableWindows("127.0.0.1", port)
+                .fail(result.reject)
+                .done(function (response) {
+                    var page = response[0];
+                    if (!page || !page.webSocketDebuggerUrl) {
+                        result.reject();
+                        return;
+                    }
+                    var _socket = new WebSocket(page.webSocketDebuggerUrl);
+                    // Disable the cache
+                    _socket.onopen = function _onConnect() {
+                        _socket.send(JSON.stringify({ id: 1, method: "Network.setCacheDisabled", params: { "cacheDisabled": true } }));
+                    };
+                    // The first message will be the confirmation => disconnected to allow remote debugging of Brackets
+                    _socket.onmessage = function _onMessage(e) {
+                        _socket.close();
+                        result.resolve();
+                    };
+                    // In case of an error
+                    _socket.onerror = result.reject;
+                });
+        }
+         
+        return result.promise();
+    }
+        
+    /**
+    * Does a full reload of the browser window
+    * @param {string} href The url to reload into the window
+    */
+    function browserReload(href) {
+        return CommandManager.execute(Commands.FILE_CLOSE_ALL, { promptOnly: true }).done(function () {
+            // Give everyone a chance to save their state - but don't let any problems block
+            // us from quitting
+            try {
+                $(ProjectManager).triggerHandler("beforeAppClose");
+            } catch (ex) {
+                console.error(ex);
+            }
+           
+            // Disable the cache to make reloads work
+            _disableCache().always(function () {
+                // Remove all menus to assure every part of Brackets is reloaded
+                _.forEach(Menus.getAllMenus(), function (value, key) {
+                    Menus.removeMenu(key);
+                });
+                
+                window.location.href = href;
+            });
+        });
+    }
+    
+    function handleReload() {
+        var href    = window.location.href,
+            params  = new UrlParams();
+        
+        // Make sure the Reload Without User Extensions parameter is removed
+        params.parse();
+        
+        if (params.get("reloadWithoutUserExts")) {
+            params.remove("reloadWithoutUserExts");
+        }
+        
+        if (href.indexOf("?") !== -1) {
+            href = href.substring(0, href.indexOf("?"));
+        }
+        
+        if (!params.isEmpty()) {
+            href += "?" + params.toString();
+        }
+        
+        // Give Mac native menus extra time to update shortcut highlighting.
+        // Prevents the menu highlighting from getting messed up after reload.
+        window.setTimeout(function () {
+            browserReload(href);
+        }, 100);
+    }
+    
+    function handleReloadWithoutExts() {
+        var href    = window.location.href,
+            params  = new UrlParams();
+        
+        params.parse();
+        
+        if (!params.get("reloadWithoutUserExts")) {
+            params.put("reloadWithoutUserExts", true);
+        }
+        
+        if (href.indexOf("?") !== -1) {
+            href = href.substring(0, href.indexOf("?"));
+        }
+        
+        href += "?" + params.toString();
+        
+        // Give Mac native menus extra time to update shortcut highlighting.
+        // Prevents the menu highlighting from getting messed up after reload.
+        window.setTimeout(function () {
+            browserReload(href);
+        }, 100);
+    }
+    
     AppInit.htmlReady(function () {
+        // If in Reload Without User Extensions mode, update UI and log console message
+        var USER_EXT_STATUS_ID = "status-user-exts";
+        
+        var params      = new UrlParams(),
+            $icon       = $("#toolbar-extension-manager"),
+            $indicator  = $("<div>" + Strings.STATUSBAR_USER_EXTENSIONS_DISABLED + "</div>");
+        
+        params.parse();
+        
+        if (params.get("reloadWithoutUserExts") === "true") {
+            CommandManager.get(Commands.FILE_EXTENSION_MANAGER).setEnabled(false);
+            $icon.css({display: "none"});
+            StatusBar.addIndicator(USER_EXT_STATUS_ID, $indicator, true);
+            console.log("Brackets reloaded with extensions disabled");
+        } else {
+            CommandManager.get(Commands.FILE_EXTENSION_MANAGER).setEnabled(true);
+            // Toolbar and status bar reload back to default states, no need to set
+        }
+        
+        // Init DOM elements
         _$titleContainerToolbar = $("#titlebar");
         _$titleWrapper = $(".title-wrapper", _$titleContainerToolbar);
         _$title = $(".title", _$titleWrapper);
         _$dirtydot = $(".dirty-dot", _$titleWrapper);
-        
     });
 
     // Exported for unit testing only
@@ -1414,10 +1550,12 @@ define(function (require, exports, module) {
     CommandManager.register(Strings.CMD_SHOW_IN_TREE,       Commands.NAVIGATE_SHOW_IN_FILE_TREE, handleShowInTree);
     CommandManager.register(Strings.CMD_SHOW_IN_OS,         Commands.NAVIGATE_SHOW_IN_OS, handleShowInOS);
     
-    // Those commands have no UI representation, and are only used internally 
-    CommandManager.registerInternal(Commands.APP_ABORT_QUIT,        _handleAbortQuit);
-    CommandManager.registerInternal(Commands.APP_BEFORE_MENUPOPUP,  _handleBeforeMenuPopup);
-    CommandManager.registerInternal(Commands.FILE_CLOSE_WINDOW,     handleFileCloseWindow);
+    // These commands have no UI representation and are only used internally
+    CommandManager.registerInternal(Commands.APP_ABORT_QUIT,            handleAbortQuit);
+    CommandManager.registerInternal(Commands.APP_BEFORE_MENUPOPUP,      handleBeforeMenuPopup);
+    CommandManager.registerInternal(Commands.FILE_CLOSE_WINDOW,         handleFileCloseWindow);
+    CommandManager.registerInternal(Commands.APP_RELOAD,                handleReload);
+    CommandManager.registerInternal(Commands.APP_RELOAD_WITHOUT_EXTS,   handleReloadWithoutExts);
     
     // Listen for changes that require updating the editor titlebar
     $(DocumentManager).on("dirtyFlagChange", handleDirtyChange);
