+++ src/extensions/default/DebugCommands/main.js
@@ -32,27 +32,19 @@ define(function (require, exports, module) {
     
     var Commands               = brackets.getModule("command/Commands"),
         CommandManager         = brackets.getModule("command/CommandManager"),
-        KeyBindingManager      = brackets.getModule("command/KeyBindingManager"),
         Menus                  = brackets.getModule("command/Menus"),
-        Editor                 = brackets.getModule("editor/Editor").Editor,
         FileSystem             = brackets.getModule("filesystem/FileSystem"),
         FileUtils              = brackets.getModule("file/FileUtils"),
-        ProjectManager         = brackets.getModule("project/ProjectManager"),
         PerfUtils              = brackets.getModule("utils/PerfUtils"),
-        NativeApp              = brackets.getModule("utils/NativeApp"),
         StringUtils            = brackets.getModule("utils/StringUtils"),
         Dialogs                = brackets.getModule("widgets/Dialogs"),
         Strings                = brackets.getModule("strings"),
-        AppInit                = brackets.getModule("utils/AppInit"),
-        UrlParams              = brackets.getModule("utils/UrlParams").UrlParams,
-        StatusBar              = brackets.getModule("widgets/StatusBar"),
         NodeDebugUtils         = require("NodeDebugUtils"),
         PerfDialogTemplate     = require("text!htmlContent/perf-dialog.html"),
         LanguageDialogTemplate = require("text!htmlContent/language-dialog.html");
     
     var KeyboardPrefs = JSON.parse(require("text!keyboard.json"));
-	
-	
+    
     /** @const {string} Brackets Application Menu Constant */
     var DEBUG_MENU = "debug-menu";
     
@@ -67,8 +59,6 @@ define(function (require, exports, module) {
         DEBUG_ENABLE_NODE_DEBUGGER      = "debug.enableNodeDebugger",
         DEBUG_LOG_NODE_STATE            = "debug.logNodeState",
         DEBUG_RESTART_NODE              = "debug.restartNode";
-    
-    
     
     function handleShowDeveloperTools() {
         brackets.app.showDeveloperTools();
@@ -89,124 +79,19 @@ define(function (require, exports, module) {
                 _testWindow = null;  // the window was probably closed
             }
         }
-
+        
         if (!_testWindow) {
             _testWindow = window.open("../test/SpecRunner.html" + queryString, "brackets-test", "width=" + $(window).width() + ",height=" + $(window).height());
             _testWindow.location.reload(true); // if it was opened before, we need to reload because it will be cached
         }
     }
     
-    /**
-     * Disables Brackets' cache via the remote debugging protocol.
-     * @return {$.Promise} A jQuery promise that will be resolved when the cache is disabled and be rejected in any other case
-     */
-    function disableCache() {
-        var result = new $.Deferred();
-        
-        if (brackets.inBrowser) {
-            result.resolve();
-        } else {
-            var Inspector = brackets.getModule("LiveDevelopment/Inspector/Inspector");
-            var port = brackets.app.getRemoteDebuggingPort ? brackets.app.getRemoteDebuggingPort() : 9234;
-            Inspector.getDebuggableWindows("127.0.0.1", port)
-                .fail(result.reject)
-                .done(function (response) {
-                    var page = response[0];
-                    if (!page || !page.webSocketDebuggerUrl) {
-                        result.reject();
-                        return;
-                    }
-                    var _socket = new WebSocket(page.webSocketDebuggerUrl);
-                    // Disable the cache
-                    _socket.onopen = function _onConnect() {
-                        _socket.send(JSON.stringify({ id: 1, method: "Network.setCacheDisabled", params: { "cacheDisabled": true } }));
-                    };
-                    // The first message will be the confirmation => disconnected to allow remote debugging of Brackets
-                    _socket.onmessage = function _onMessage(e) {
-                        _socket.close();
-                        result.resolve();
-                    };
-                    // In case of an error
-                    _socket.onerror = result.reject;
-                });
-        }
-         
-        return result.promise();
-    }
-        
-    /**
-    * Does a full reload of the browser window
-    * @param {string} href The url to reload into the window
-    */
-    function browserReload(href) {
-        return CommandManager.execute(Commands.FILE_CLOSE_ALL, { promptOnly: true }).done(function () {
-            // Give everyone a chance to save their state - but don't let any problems block
-            // us from quitting
-            try {
-                $(ProjectManager).triggerHandler("beforeAppClose");
-            } catch (ex) {
-                console.error(ex);
-            }
-           
-            // Disable the cache to make reloads work
-            disableCache().always(function () {
-                // Remove all menus to assure every part of Brackets is reloaded
-                _.forEach(Menus.getAllMenus(), function (value, key) {
-                    Menus.removeMenu(key);
-                });
-                
-                window.location.href = href;
-            });
-        });
-    }
-    
     function handleReload() {
-        var href    = window.location.href,
-            params  = new UrlParams();
-        
-        // Make sure the Reload Without User Extensions parameter is removed
-        params.parse();
-        
-        if (params.get("reloadWithoutUserExts")) {
-            params.remove("reloadWithoutUserExts");
-        }
-        
-        if (href.indexOf("?") !== -1) {
-            href = href.substring(0, href.indexOf("?"));
-        }
-        
-        if (!params.isEmpty()) {
-            href += "?" + params.toString();
-        }
-        
-        // Give Mac native menus extra time to update shortcut highlighting.
-        // Prevents the menu highlighting from getting messed up after reload.
-        window.setTimeout(function () {
-            browserReload(href);
-        }, 100);
+        CommandManager.execute(Commands.APP_RELOAD);
     }
     
     function handleReloadWithoutUserExts() {
-        var href    = window.location.href,
-            params  = new UrlParams();
-        
-        params.parse();
-        
-        if (!params.get("reloadWithoutUserExts")) {
-            params.put("reloadWithoutUserExts", true);
-        }
-        
-        if (href.indexOf("?") !== -1) {
-            href = href.substring(0, href.indexOf("?"));
-        }
-        
-        href += "?" + params.toString();
-        
-        // Give Mac native menus extra time to update shortcut highlighting.
-        // Prevents the menu highlighting from getting messed up after reload.
-        window.setTimeout(function () {
-            browserReload(href);
-        }, 100);
+        CommandManager.execute(Commands.APP_RELOAD_WITHOUT_EXTS);
     }
         
     function handleNewBracketsWindow() {
@@ -310,7 +195,7 @@ define(function (require, exports, module) {
                 Dialogs.showModalDialogUsingTemplate(template).done(function (id) {
                     if (id === Dialogs.DIALOG_BTN_OK && locale !== curLocale) {
                         brackets.setLocale(locale);
-                        CommandManager.execute(DEBUG_REFRESH_WINDOW);
+                        CommandManager.execute(Commands.APP_RELOAD);
                     }
                 });
                 
@@ -366,7 +251,6 @@ define(function (require, exports, module) {
     
     enableRunTestsMenuItem();
     
-    
     /*
      * Debug menu
      */
@@ -385,28 +269,6 @@ define(function (require, exports, module) {
     menu.addMenuItem(DEBUG_LOG_NODE_STATE);
     menu.addMenuItem(DEBUG_RESTART_NODE);
     
-    
     // exposed for convenience, but not official API
     exports._runUnitTests = _runUnitTests;
-    
-    AppInit.htmlReady(function () {
-        // If in Reload Without User Extensions mode, update UI and log console message
-        var USER_EXT_STATUS_ID = "status-user-exts";
-        
-        var params      = new UrlParams(),
-            $icon       = $("#toolbar-extension-manager"),
-            $indicator  = $("<div>" + Strings.STATUSBAR_USER_EXTENSIONS_DISABLED + "</div>");
-        
-        params.parse();
-        
-        if (params.get("reloadWithoutUserExts") === "true") {
-            CommandManager.get(Commands.FILE_EXTENSION_MANAGER).setEnabled(false);
-            $icon.css({display: "none"});
-            StatusBar.addIndicator(USER_EXT_STATUS_ID, $indicator, true);
-            console.log("Brackets reloaded with extensions disabled");
-        } else {
-            CommandManager.get(Commands.FILE_EXTENSION_MANAGER).setEnabled(true);
-            // Toolbar and status bar reload back to default states, no need to set
-        }
-    });
 });
